requirements.txt
elevenlabs==0.2.27
python-dotenv==1.0.0
pydub==0.25.1
pyaudio==0.2.13 